requirements.txt
opencv-python~=4.9.0.80
aiohttp~=3.9.5
aiortc~=1.8.0
#av~=11.0.0
torch~=2.0.0
albumentations~=1.3.1
numpy~=1.21.6
torchmetrics~=1.4.0.post0

ultralytics==8.0.88
av~=11.0.0
PyYAML~=6.0.1
